move around folders so they make sense
