--- onevoice/src/App.jsx
@@ -1,14 +1,0 @@
-import { useState } from "react";
-
-import "./App.css";
-import { TranscriptionComponent } from "./components/AssemblyAi.jsx";
-
-function App() {
-  return (
-    <>
-      <TranscriptionComponent />
-    </>
-  );
-}
-
-export default App;
